# api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Pedidos",
        "description": "API para gestionar pedidos y ventas, incluyendo la creación de pedidos, consulta de productos y entrega.",
        "contact": {
            "email": "soporte@tusistema.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "api/administradores/productos/registrar": {
            "post": {
                "tags": [
                    "Productos"
                ],
                "summary": "Crear un nuevo producto",
                "operationId": "3a779db3dc2f9164c5059572dba4ddc2",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre",
                                    "precio_venta",
                                    "precio_base",
                                    "medida"
                                ],
                                "properties": {
                                    "nombre": {
                                        "description": "Nombre del producto",
                                        "type": "string",
                                        "example": "Producto A"
                                    },
                                    "descripcion": {
                                        "description": "Descripción del producto",
                                        "type": "string",
                                        "example": "Descripción detallada del Producto A"
                                    },
                                    "precio_venta": {
                                        "description": "Precio de venta",
                                        "type": "number",
                                        "format": "float",
                                        "example": 100.5
                                    },
                                    "precio_base": {
                                        "description": "Precio base",
                                        "type": "number",
                                        "format": "float",
                                        "example": 80
                                    },
                                    "medida": {
                                        "description": "Unidad de medida",
                                        "type": "string",
                                        "enum": [
                                            "unidades",
                                            "metros",
                                            "kilos",
                                            "litros"
                                        ],
                                        "example": "unidades"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Producto creado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Producto"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación"
                    },
                    "500": {
                        "description": "Error en la base de datos"
                    }
                }
            }
        },
        "api/administradores/productos/listar": {
            "get": {
                "tags": [
                    "Productos"
                ],
                "summary": "Listar todos los productos",
                "operationId": "0df045a3deaba8e493cfcce1c6d9ef05",
                "responses": {
                    "200": {
                        "description": "Lista de productos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Producto"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error al obtener productos"
                    }
                }
            }
        },
        "api/administradores/productos/buscar/{id}": {
            "get": {
                "tags": [
                    "Productos"
                ],
                "summary": "Buscar un producto por ID",
                "operationId": "7d69078d9de2d8678dd3448d7ad3ef4a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del producto a obtener",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Datos del producto",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Producto"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Producto no encontrado"
                    },
                    "500": {
                        "description": "Error al obtener el producto"
                    }
                }
            }
        },
        "api/administradores/productos/actualizar/{id}": {
            "put": {
                "tags": [
                    "Productos"
                ],
                "summary": "Actualizar un producto",
                "operationId": "b589555df4ebd3ffa86028dece766139",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del producto a actualizar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre",
                                    "precio_venta",
                                    "precio_base",
                                    "medida"
                                ],
                                "properties": {
                                    "nombre": {
                                        "description": "Nombre del producto",
                                        "type": "string",
                                        "example": "Producto A"
                                    },
                                    "descripcion": {
                                        "description": "Descripción del producto",
                                        "type": "string",
                                        "example": "Descripción detallada del Producto A"
                                    },
                                    "precio_venta": {
                                        "description": "Precio de venta",
                                        "type": "number",
                                        "format": "float",
                                        "example": 100.5
                                    },
                                    "precio_base": {
                                        "description": "Precio base",
                                        "type": "number",
                                        "format": "float",
                                        "example": 80
                                    },
                                    "medida": {
                                        "description": "Unidad de medida",
                                        "type": "string",
                                        "enum": [
                                            "unidades",
                                            "metros",
                                            "kilos",
                                            "litros"
                                        ],
                                        "example": "unidades"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Producto actualizado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Producto"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Producto no encontrado"
                    },
                    "500": {
                        "description": "Error al actualizar el producto"
                    }
                }
            }
        },
        "api/administradores/productos/eliminar/{id}": {
            "delete": {
                "tags": [
                    "Productos"
                ],
                "summary": "Eliminar un producto",
                "operationId": "b6e97afb83ae5a1f68c2e6d5af703cf8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del producto a eliminar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Producto eliminado"
                    },
                    "404": {
                        "description": "Producto no encontrado"
                    },
                    "500": {
                        "description": "Error al eliminar el producto"
                    }
                }
            }
        },
        "api/administradores/productos/add/stock/{id}": {
            "post": {
                "tags": [
                    "Productos"
                ],
                "summary": "Añadir stock a un producto",
                "operationId": "1f9d67924aa686c5c6807a2d95bb5d8a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del producto al que se le añadirá stock",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "cantidad"
                                ],
                                "properties": {
                                    "cantidad": {
                                        "description": "Cantidad de stock a añadir",
                                        "type": "number",
                                        "format": "float",
                                        "example": 10
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Stock añadido exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Producto"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Producto no encontrado"
                    },
                    "500": {
                        "description": "Error al añadir el stock"
                    }
                }
            }
        },
        "api/administradores/productos/remove/stock/{id}": {
            "post": {
                "tags": [
                    "Productos"
                ],
                "summary": "Quitar stock a un producto",
                "operationId": "06fbfe7e34d9a3048b73d75852828b0f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del producto del que se le quitará stock",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "cantidad"
                                ],
                                "properties": {
                                    "cantidad": {
                                        "description": "Cantidad de stock a quitar",
                                        "type": "number",
                                        "format": "float",
                                        "example": 5
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Stock quitado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Producto"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Producto no encontrado"
                    },
                    "500": {
                        "description": "Error al quitar el stock"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Producto": {
                "required": [
                    "id",
                    "estado",
                    "fecha_creacion",
                    "nombre",
                    "precio_venta",
                    "precio_base",
                    "medida"
                ],
                "properties": {
                    "id": {
                        "description": "ID del producto",
                        "type": "integer",
                        "example": 101
                    },
                    "estado": {
                        "description": "Estado del producto",
                        "type": "string",
                        "enum": [
                            "ACTIVO",
                            "INACTIVO"
                        ],
                        "example": "ACTIVO"
                    },
                    "fecha_creacion": {
                        "description": "Fecha de creación del producto",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-11-27T10:00:00Z"
                    },
                    "fecha_actualizacion": {
                        "description": "Fecha de actualización del producto",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-11-27T12:00:00Z"
                    },
                    "nombre": {
                        "description": "Nombre del producto",
                        "type": "string",
                        "example": "Producto A"
                    },
                    "descripcion": {
                        "description": "Descripción del producto",
                        "type": "string",
                        "example": "Este es un producto de ejemplo."
                    },
                    "precio_venta": {
                        "description": "Precio de venta del producto",
                        "type": "number",
                        "format": "float",
                        "example": 20.5
                    },
                    "precio_base": {
                        "description": "Precio base del producto",
                        "type": "number",
                        "format": "float",
                        "example": 15
                    },
                    "medida": {
                        "description": "Unidad de medida del producto",
                        "type": "string",
                        "enum": [
                            "unidades",
                            "metros",
                            "kilos",
                            "litros"
                        ],
                        "example": "kilos"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Utiliza un token de autorización Bearer para autenticar las solicitudes.",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}